Fix Rubocop offenses (again)
Split history tests into two separate files
